// Angular/Management/src/app/model/Staff.ts
export class Staff{
    staffName: string="";
    staffNo: number=0;
    staffAdd: string="";
    staffQuali: string="";
    staffSub: string="";
    staffJoin: string="";
    staffMail: string="";
    id: number=0;
    staffBirth: string="";
    staffExperience: number=0;
    staffCertificate: string="";
    staffSalary: number=0;
}